pyproject.toml
[project]
name = "bcd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "isort>=6.0.1",
    "langchain-mcp-adapters>=0.1.1",
    "langchain[anthropic]>=0.3.25",
    "langgraph>=0.4.7",
    "python-dotenv>=1.1.0",
]
